put geticontextcolor into profile functions
