+++ src/Components/Navbar/UserNavbar.js
@@ -4,6 +4,7 @@ import { useUser } from '../context/UserContext';
 import { useAuth } from '../context/AuthContext';
 import { getUserById } from '../../APIFunctions/User';
 import { useBackgroundColor } from '../context/BackgroundColorContext';
+import { getIconTextColor } from '../../APIFunctions/Profile';
 
 export default function UserNavbar(props) {
   const { user } = useUser();
@@ -36,35 +37,6 @@ export default function UserNavbar(props) {
       }
     };
   }, [authenticated, user, backgroundColorVersion]);
-
-  // using w3c guidelines
-  function getIconTextColor(color) {
-    if(typeof color !== 'string') {
-      throw new TypeError('color must be a string');
-    }
-    if(color == '' || color == '#2a323c') {
-      return '#FFFFFF';
-    }
-    // get rgb values 0-255
-    const r = parseInt(color.substring(1, 3), 16);
-    const g = parseInt(color.substring(3, 5), 16);
-    const b = parseInt(color.substring(5, 7), 16);
-    // linearize the colors
-    const colors = [r / 255.0, g / 255.0, b / 255.0];
-    const c = colors.map((color) => {
-      if(color <= 0.04045) {
-        return color / 12.92;
-      }
-      return Math.pow(((color + 0.055) / 1.055), 2.4);
-    });
-    // luminance value
-    const L = 0.2126 * c[0] + 0.7152 * c[1] + 0.0722 * c[2];
-    // threshold of 0.179
-    if(L > 0.179) {
-      return '#000000';
-    }
-    return '#FFFFFF';
-  }
 
   let initials = '';
   if (user && user.firstName && user.lastName) {
